requirements.txt
requests>=2.25.0
pysha3>=1.0.2; python_version<"3.6"
